GigaExchange/requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
backtrader==1.9.76.123
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.3
click==8.0.1
cycler==0.10.0
dateparser==1.0.0
Flask==2.0.1
gunicorn==20.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
multidict==5.1.0
numpy==1.21.1
Pillow==8.3.1
pyparsing==2.4.7
python-binance==1.0.12
python-dateutil==2.8.2
pytz==2021.1
regex==2021.7.6
requests==2.26.0
six==1.16.0
typing-extensions==3.10.0.0
tzlocal==2.1
ujson==4.0.2
urllib3==1.26.6
websockets==9.1
Werkzeug==2.0.1
yarl==1.6.3